+++ docs/changes.md
@@ -5,6 +5,12 @@ title: Support Change Log
 
 ## Version 3.3 {#v3-3}
 
+### v3.3.2 {#v3-3-2}
+
+* Add `Orchestra\Support\Arr::expand()` method. ([@mustafaaloko](https://github.com/mustafaaloko))
+* Add `Orchestra\Support\Serializer`.
+* Add `Orchestra\Support\Transformer`.
+
 ### v3.3.1 {#v3-3-1}
 
 * Core: